Fix Accessory decoder packet routine
The accessory code was computing the wrong address. Accessory decoder addressing is not standard and causes much confusion. What the NMRA spec says and what most people and base stations use is different. While this is true:

10AAAAAA 0 1AAACDDD

The way it is implemented varies. The spec says that the A's in the first byte are the least significant bits of the address and the A's in the second byte are the 1s complement most significant bits that make up a 9 bit address. C is supposed to be "activate" and "deactivate" and some rare equipment uses this. However, the standard method is to have 9 bits for the address and bits 1 and 2 of byte 2 make up a "subaddress" representing the 4 outputs of many turnouts. Bit 0 becomes the "closed/thrown" switch and is basically the "on/off" instead of C. NMRA even says that "C should always be set to 1", despite instruction elsewhere to the contrary. 

I changed the code to compute the correct 2 part (address, subaddress) pair and added the correct 11bit (the 9 bit address with the sub address) address for those devices/programs that use straight linear addressing. An example of accessory decoder addressing:

linear address / 2 part address
1273 / 319,0
1274 / 319,1
1275 / 319,2
1276 / 319,3

At 1277 it starts over again at port 0. In DCC++ you would enter <a 319 3 1> to "throw" or turn on the accessory at address 1276 aka address 319 output 3. Then <a 319 3 0> closes it.  Bit 0 in the second byte (the last of the DDD bits) actually acts as the toggle to open and close the switch. JMRI can handle this automatically and you can just enter 1276 instead of the 2 part address.

The question we may want to answer at some point is how to deal with people who may want to see that last bit as the "on" and "off" switch not the C position in bit 1. For now, I Added the 11 bit linear address, fixed the 2 part address, and added the last bit so people using the normal addressing can still see on and off by looking at that bit while others can look at the "on" and "off" and see that. For most people, it will always be "on" or "active" and the "closed/thrown" bit will let them know the status. So it looks like this:

address: 229 / 58,0,1 on

If more people use this, they can comment on how they would like it displayed. maybe just the bits should be displayed like:

address: 229 / 58,0 thrown:1 activate:1 or address: 229 / 58,0,1 C:1